refactor(gapi): extract google auth from calendar
API of Google Sheet and profiles will reuse code of Google authentication
